Split out translate-toolkit based formats to separate files
